fix: refactor store
