+++ apps/web/src/views/Schedule/MarksByDay/LessonGrades.tsx
@@ -1,13 +1,13 @@
 import { Mark } from '@components'
 import { MODAL_PAGE_MARK } from '@config'
 import { Task } from '@diary-spo/shared'
+import { useMarkModal } from '@store'
 import { setDefaultMark, truncateString } from '@utils'
 import { useRouteNavigator } from '@vkontakte/vk-mini-apps-router'
 import { Header, HorizontalCell } from '@vkontakte/vkui'
 import { FunctionalComponent } from 'preact'
 import { CSSProperties } from 'preact/compat'
 import { useCallback } from 'preact/hooks'
-import useMarkModal from '../../../store/useMarkModal.ts'
 import { LessonGradesProps } from './types.ts'
 
 const marksGap: CSSProperties = { display: 'flex' }
